+++ Examples/InstreamAdExample/app/src/main/java/com/yandex/instream/sample/player/SamplePlayer.java
@@ -1,0 +1,18 @@
+/*
+ * This file is a part of the Yandex Advertising Network
+ *
+ * Version for Android (C) 2021 YANDEX
+ *
+ * You may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at https://legal.yandex.com/partner_ch/
+ */
+package com.yandex.instream.sample.player;
+
+public interface SamplePlayer {
+
+    boolean isPlaying();
+
+    void onResume();
+
+    void onPause();
+}